A package for classifying border crossings using a rules-based methodology.

The migrbc package provides three categories of important functions: run_rbc, pre_process, plot_mig_hist, resolve_data, initialize_logger, and resolve_data_with_error. In among of the five functions, run_rbc is the main entry function of the package. Three functions: initialize_logger, pre_process and plot_mig_hit are utility functions. The rest functions resolve_data and resolve_data_with_error are the key functions to do the rules based classification.

initialize_logger function

This function is used to initialize the futile.logger so that the user can be notified with the current status of running RBC.

run_rbc function

The run_rbc function attempt to determine long-term migration statuses, and pre-crossing and post-crossing residence statuses, for all crossings where these statuses are not known.

pre_process function

This function provides a mechanism to divide large data into small chunks.

plot_mig_hist function

Given a sequence of border crossings for a person, draw a diagram describing that person's migration history

resolve_data function

This function is the key function to do the rules based classification.

resolve_data_with_error function

This function is used to produce error result.