Processing RBC for a person.

This function is used to resolve one person's journeys, i.e., classifying a person and marking it whether or not to be a long term migrant based on the person's journeys. This function is used internally inside the package and shouldn't be exposed to the outside caller.

rcpp_resolve(person_data, int_res_status, initial_date_finalised, tw, tm)

Arguments

person_data
A list object of a person's journeys.

int_res_status
The initial residence status of the target person

initial_date_finalised
The final resolved date of the initial residence status.

tw
Windows Size, by default, it is 487 days.

tm
Threshold of Year, by default, it is 365 days.

Value

A list of classified / labelled journeys.