Apply rule-based classification to identify which migration or residence status are unknown

Attempt to determine long-term migration statuses, and pre-crossing and post-crossing residence statuses, for all crossings where these statuses are not known.

run_rbc(crossing_data, init_res_status_data = NULL, window_size = 487,
  threshold_year = 365, parallel = FALSE, n_core = 2, max_ram = 2,
  include_error_columns = FALSE, mc.cleanup = FALSE)

Arguments

crossing_data
A pre-processed group data contain journeys, movements and IRS or the raw crossing data.

init_res_status_data
The raw data of the initial residence status in the format of data frame.

window_size
The maximum length of the scanning period. Can be an integer giving the number of days, the result of a call to function difftime, or an object of class Duration.

threshold_year
The length of the yearly test period. It can be an integer giving the number of days, the result of a call to function difftime, or an object of class Duration.

parallel
Logical. Whether to use parallel processing, to speed up the calculation of migration statuses. Defaults to TRUE.

n_core
The number of cores to use, if parallel is TRUE. Defaults to 2. Higher values will typically result in faster calculations on computers with more than two cores.

max_ram
Optional, it is used to limit the RAM that can be used by this function. The default value is 2 Gb.

include_error_columns
Optional, if it is TRUE, the returned result of error_data will contain two extra columns error_code and error_message.

mc.cleanup
Optional, if set to TRUE then all children that have been forked by this function will be killed (by sending SIGTERM) before this function returns. Under normal circumstances mclapply waits for the children to deliver results, so this option usually has only effect when mclapply is interrupted. If set to FALSE then child processes are collected, but not forcefully terminated. As a special case this argument can be set to the number of the signal that should be used to kill the children instead of SIGTERM.

Value

The number of rows in the FinalIsLongTermMig table.

Examples

## generate test data 100 people and each person has 
## 10 journeys
number_of_people <- 100
person_data <- migrbc::setup_random_test_data(
    number_of_people,
    initial_date = '2001-01-01',
    numJourneys = 10,
    min = 0,
    max = 100)
head(person_data)

#>   journeyId personId is_arrival date_crossing journey_sequence
#> 1         1        1          1    2001-01-01                1
#> 2         2        1          0    2001-01-11                2
#> 3         3        1          1    2001-01-22                3
#> 4         4        1          0    2001-02-13                4
#> 5         5        1          1    2001-03-15                5
#> 6         6        1          0    2001-04-26                6
#>   res_status_before res_status_after journeyId_prev
#> 1                 0                0              0
#> 2                 0                0              0
#> 3                 0                0              0
#> 4                 0                0              0
#> 5                 0                0              0
#> 6                 0                0              0

cross_spaces <- migrbc::pre_process(person_data)

#> INFO [2020-05-01 13:05:47] entering 'pre_process' :
#> INFO [2020-05-01 13:05:47] leaving 'pre_process' : 
#>               return 1 groups of raw data.

## run in non-parallel
res <- migrbc::run_rbc(cross_spaces,
                       window_size = 487,
                       threshold_year = 365,
                       parallel=FALSE)

#> INFO [2020-05-01 13:05:47] entering 'run_rbc' :
#>                            window_size = 487, 
#>                            threshold_year = 365, 
#>                            parallel = FALSE, 
#>                            n_core = 2,
#>                            max_ram = 2,
#>                            include_error_columns = FALSE,
#>                            mc.cleanup = FALSE
#> INFO [2020-05-01 13:05:47] entering 'resolve_data' :
#> INFO [2020-05-01 13:05:47] Running without parallel
#> INFO [2020-05-01 13:05:47] entering 'resolve_data, chunk =1':
#>                window_size=487, 
#>                threshold_year=365
#> INFO [2020-05-01 13:05:48] leaving 'resolve_data' :
#> INFO [2020-05-01 13:05:48] leaving 'run_rbc'

## run in parallel with n_core = 2
cross_spaces <- migrbc::pre_process(person_data, n_groups = 2)

#> INFO [2020-05-01 13:05:48] entering 'pre_process' :
#> INFO [2020-05-01 13:05:48] leaving 'pre_process' : 
#>               return 2 groups of raw data.

res <- migrbc::run_rbc(cross_spaces,
                       window_size = 487,
                       threshold_year = 365,
                       parallel=TRUE,
                       n_core = 2)

#> INFO [2020-05-01 13:05:48] entering 'run_rbc' :
#>                            window_size = 487, 
#>                            threshold_year = 365, 
#>                            parallel = TRUE, 
#>                            n_core = 2,
#>                            max_ram = 2,
#>                            include_error_columns = FALSE,
#>                            mc.cleanup = FALSE
#> INFO [2020-05-01 13:05:48] entering 'resolve_data' :
#> INFO [2020-05-01 13:05:48] Running with parallel:
#>                  n_cores = 2
#> INFO [2020-05-01 13:05:49] leaving 'resolve_data' :
#> INFO [2020-05-01 13:05:49] leaving 'run_rbc'

head(res$journeys)

#>   journeyId journeyId_prev personId date_crossing journey_sequence
#> 1         1              0        1    2001-01-01                1
#> 2         2              1        1    2001-01-11                2
#> 3         3              2        1    2001-01-22                3
#> 4         4              3        1    2001-02-13                4
#> 5         5              4        1    2001-03-15                5
#> 6         6              5        1    2001-04-26                6
#>   days_to_next_crossing is_arrival res_status_before res_status_after
#> 1                    10          1                 0                0
#> 2                    11          0                 0                0
#> 3                    22          1                 0                0
#> 4                    30          0                 0                0
#> 5                    42          1                 0                0
#> 6                    47          0                 0                0
#>   is_long_term_mig date_finalised_res_before date_finalised_res_after
#> 1                0                2001-01-01               2001-09-22
#> 2                0                2001-09-22               2001-09-22
#> 3                0                2001-09-22               2001-10-03
#> 4                0                2001-10-03               2001-10-03
#> 5                0                2001-10-03               2002-02-04
#> 6                0                2002-02-04               2002-02-04
#>   date_finalised_LTM
#> 1         2001-09-22
#> 2         2001-01-11
#> 3         2001-10-03
#> 4         2001-02-13
#> 5         2002-02-04
#> 6         2001-04-26

head(res$error_data)

#>  [1] journeyId                 journeyId_prev           
#>  [3] personId                  date_crossing            
#>  [5] journey_sequence          days_to_next_crossing    
#>  [7] is_arrival                res_status_before        
#>  [9] res_status_after          is_long_term_mig         
#> [11] date_finalised_res_before date_finalised_res_after 
#> [13] date_finalised_LTM       
#> <0 rows> (or 0-length row.names)